dealbrain/pyproject.toml
[project]
name = "dealbrain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "google-adk>=1.28.1",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-httplib2>=0.3.1",
    "google-cloud-storage>=3.10.1",
    "google-genai>=1.70.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.43.0",
]
